Create LazyCheckRootNode for the whole TypeCheckValueNode.check node tree
